deliverable 5- create delete api

+++ controllers/user.js
@@ -37,17 +37,29 @@ const getuser = async (req, res) => {
 
 const deleteuser = async (req, res) => {
   try {
-    // if (!req.params.id == "undefined") {
-    //   console.log("Id is missing");
-    //   return res.status(400).json({ err: "Id is missing" });
-    // }
+    if (req.param.id === "undefined") {
+      console.log("Id is missing");
+      return res.status(400).json({ err: "Id is missing" });
+    }
+
     const userId = req.params.id;
     await User.destroy({ where: { id: userId } });
-    return res.Status(200);
+    return res.status(200);
   } catch (err) {
     console.log("delete", err);
     res.status(500).json(err);
   }
+
+  // const uId = req.body.userId;
+  // User.findByPk(uId)
+  //   .then((user) => {
+  //     User.destroy({ where: { id: uId } });
+  //   })
+  //   .then((result) => {
+  //     console.log("DESTROYED USER");
+  //     res.redirect("/");
+  //   })
+  //   .catch((err) => console.log(err));
 };
 
 module.exports = { adduser, getuser, deleteuser };
